+++ src/test/java/com/fasterxml/jackson/databind/struct/TestObjectIdDeserialization.java
@@ -1,23 +1,20 @@
 package com.fasterxml.jackson.databind.struct;
 
-import java.util.ArrayList;
-import java.util.EnumMap;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.concurrent.ArrayBlockingQueue;
 
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
-import com.fasterxml.jackson.databind.*;
+import com.fasterxml.jackson.databind.BaseMapTest;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.deser.UnresolvedForwardReference;
 import com.fasterxml.jackson.databind.deser.UnresolvedForwardReference.UnresolvedId;
 import com.fasterxml.jackson.databind.struct.TestObjectId.Company;
 import com.fasterxml.jackson.databind.struct.TestObjectId.Employee;
-import com.fasterxml.jackson.databind.struct.TestObjectIdDeserialization.EnumMapCompany.FooEnum;
 
 /**
  * Unit test to verify handling of Object Id deserialization
@@ -118,34 +115,6 @@ public class TestObjectIdDeserialization extends BaseMapTest
     
     static class MappedCompany {
         public Map<Integer, Employee> employees;
-    }
-
-    static class ArrayCompany {
-        public Employee[] employees;
-    }
-
-    static class ArrayBlockingQueueCompany {
-        public ArrayBlockingQueue<Employee> employees;
-    }
-
-    static class EnumMapCompany {
-        public EnumMap<FooEnum,Employee> employees;
-
-        static enum FooEnum {
-            A, B
-        }
-    }
-
-    static class DefensiveCompany {
-        public List<DefensiveEmployee> employees;
-
-        static class DefensiveEmployee extends Employee {
-
-            public void setReports(List<DefensiveEmployee> reports)
-            {
-                this.reports = new ArrayList<TestObjectId.Employee>(reports);
-            }
-        }
     }
 
     @JsonIdentityInfo(generator = ObjectIdGenerators.IntSequenceGenerator.class)
@@ -268,64 +237,6 @@ public class TestObjectIdDeserialization extends BaseMapTest
         assertEquals(3, company.employees.size());
         Employee firstEmployee = company.employees.get(1);
         Employee secondEmployee = company.employees.get(3);
-        assertEmployees(firstEmployee, secondEmployee);
-    }
-
-    public void testForwardReferenceInArray()
-        throws Exception
-    {
-        String json = "{\"employees\":["
-                      + "{\"id\":1,\"name\":\"First\",\"manager\":null,\"reports\":[2]},"
-                      + "2,"
-                      +"{\"id\":2,\"name\":\"Second\",\"manager\":1,\"reports\":[]}"
-                      + "]}";
-        ArrayCompany company = mapper.readValue(json, ArrayCompany.class);
-        assertEquals(3, company.employees.length);
-        Employee firstEmployee = company.employees[0];
-        Employee secondEmployee = company.employees[1];
-        assertEmployees(firstEmployee, secondEmployee);
-    }
-
-    // Do a specific test for ArrayBlockingQueue since it has its own deser.
-    public void testForwardReferenceInQueue()
-        throws Exception
-    {
-        String json = "{\"employees\":["
-                      + "{\"id\":1,\"name\":\"First\",\"manager\":null,\"reports\":[2]},"
-                      + "2,"
-                      +"{\"id\":2,\"name\":\"Second\",\"manager\":1,\"reports\":[]}"
-                      + "]}";
-        ArrayBlockingQueueCompany company = mapper.readValue(json, ArrayBlockingQueueCompany.class);
-        assertEquals(3, company.employees.size());
-        Employee firstEmployee = company.employees.take();
-        Employee secondEmployee = company.employees.take();
-        assertEmployees(firstEmployee, secondEmployee);
-    }
-
-    public void testForwardReferenceInEnumMap()
-        throws Exception
-   {
-        String json = "{\"employees\":{"
-                      + "\"A\":{\"id\":1,\"name\":\"First\",\"manager\":null,\"reports\":[2]},"
-                      + "\"B\": 2,"
-                      + "\"C\":{\"id\":2,\"name\":\"Second\",\"manager\":1,\"reports\":[]}"
-                      + "}}";
-        EnumMapCompany company = mapper.readValue(json, EnumMapCompany.class);
-        assertEquals(3, company.employees.size());
-        Employee firstEmployee = company.employees.get(FooEnum.A);
-        Employee secondEmployee = company.employees.get(FooEnum.B);
-        assertEmployees(firstEmployee, secondEmployee);
-    }
-
-    public void testForwardReferenceWithDefensiveCopy()
-        throws Exception
-    {
-        String json = "{\"employees\":[" + "{\"id\":1,\"name\":\"First\",\"manager\":null,\"reports\":[2]},"
-                + "{\"id\":2,\"name\":\"Second\",\"manager\":1,\"reports\":[]}" + "]}";
-        DefensiveCompany company = mapper.readValue(json, DefensiveCompany.class);
-        assertEquals(2, company.employees.size());
-        Employee firstEmployee = company.employees.get(0);
-        Employee secondEmployee = company.employees.get(1);
         assertEmployees(firstEmployee, secondEmployee);
     }
